Module Mirage_device

Abstract devices

This module define the basic functions that a MirageOS device should implement.

Release v2.0.0

type error = [

| `Unimplemented
operation not yet implemented in the code

| `Disconnected
the device has been previously disconnected

]

The type for device errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for errors.

module type S = sig ... end

Defines the functions to define what is a device state and how to disconnect such a device.