Module type Mirage_device_lwt.S

Defines the functions to define what is a device state and how to disconnect such a device.

type t

The type representing the internal state of the device

val disconnect : t -> unit Lwt.t

Disconnect from the device. While this might take some time to complete, it can never result in an error.